% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{create_norm_miss_data}
\alias{create_norm_miss_data}
\title{create_norm_miss_data}
\usage{
create_norm_miss_data(
  prop_miss,
  alpha = 0,
  beta = 1,
  sigma2 = 1e-04,
  is_mar = TRUE,
  n = 200
)
}
\arguments{
\item{prop_miss}{Proportion of missing data -> \link{0, 1}}

\item{alpha}{y = alpha + beta * x + error, where error ~ N(0, sigma2)}

\item{beta}{y = alpha + beta * x + error, where error ~ N(0, sigma2)}

\item{sigma2}{y = alpha + beta * x + error, where error ~ N(0, sigma2)}

\item{is_mar}{Missing mechanism -> Missing at random (TRUE) or missing completely at random (FALSE)}

\item{n}{Number of observations}
}
\value{
dt
}
\description{
create_norm_miss_data
}
