% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_weights_bayes.R
\name{draw_weights_bayes}
\alias{draw_weights_bayes}
\title{draw_weights_bayes}
\usage{
draw_weights_bayes(y_obs, X_obs, beta_hat, V, n1, q)
}
\arguments{
\item{y_obs}{Observed outcomes}

\item{X_obs}{Predictors associated with observed outcomes}

\item{beta_hat}{Estimated weights based on observed data}

\item{n1}{Number of observed outcomes}

\item{q}{Number of predictors (including the intercept)}
}
\value{
beta_dot
}
\description{
draw_weights_bayes
}
