% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_weights_boot.R
\name{draw_weights_boot}
\alias{draw_weights_boot}
\title{draw_weights_boot}
\usage{
draw_weights_boot(y_obs, X_obs, beta_hat, V, n1, q, ridge)
}
\arguments{
\item{y_obs}{Observed outcomes}

\item{X_obs}{Predictors associated with observed outcomes}

\item{beta_hat}{Estimated weights based on observed data}

\item{n1}{Number of observed outcomes}

\item{q}{Number of predictors (including the intercept)}

\item{ridge}{Ridge penalty (a small number close to 0)}
}
\value{
beta_dot
}
\description{
draw_weights_boot
}
