% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{eval_pmm}
\alias{eval_pmm}
\title{eval_pmm}
\usage{
eval_pmm(simu_results, true_coef)
}
\arguments{
\item{simu_results}{Results of simulation generated by simulation()}

\item{true_coef}{The true value of coefficient to evaluate}
}
\value{
eval_metrics
}
\description{
eval_pmm
}
