% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{get_eval_results}
\alias{get_eval_results}
\title{get_eval_results}
\usage{
get_eval_results(data, true_coef, donors, matchtype, simu_times = 500)
}
\arguments{
\item{true_coef}{The true value of coefficient to evaluate}

\item{donors}{The size of donor pool (usually between 3 and 10)}

\item{matchtype}{Type of matching distance -> {0, 1, 2}}

\item{simu_times}{}
}
\value{
results
}
\description{
get_eval_results
}
