% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_matcher.R
\name{naive_matcher}
\alias{naive_matcher}
\title{naive_matcher}
\usage{
naive_matcher(y_obs_hat, y_mis_hat, n1, n0, donors)
}
\arguments{
\item{y_obs_hat}{Estimated observed outcomes using weights calculated with observed data}

\item{y_mis_hat}{Estimated missing outcomes using weights calculated with observed data}

\item{n1}{Number of observed outcomes}

\item{n0}{Number of missing outcomes/values to impute}

\item{donors}{The size of donor pool (usually between 3 and 10)}
}
\value{
imputed_ind
}
\description{
naive_matcher
}
