% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_impute.R
\name{pmm_impute}
\alias{pmm_impute}
\title{pmm_impute}
\usage{
pmm_impute(
  y,
  ry,
  x,
  wy = NULL,
  bayeswt = TRUE,
  donors = 5,
  matchtype = 1,
  c_matcher = TRUE,
  ridge = 1e-05
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Location of observed values -> (1) observed (TRUE) vs. missing (FALSE); (2) length(ry) == length(y)}

\item{x}{Numeric design matrix (no intercepts)}

\item{wy}{Location of values to impute -> (1) create imputations (TRUE) vs. not (FALSE); (2) length(wy) == length(y)}

\item{bayeswt}{Type of regression weights -> "Bayesian" (TRUE) or "Bootstrap" (FALSE)}

\item{donors}{The size of donor pool (usually between 3 and 10)}

\item{matchtype}{Type of matching distance -> {0, 1, 2}}

\item{ridge}{Ridge penalty (a small number close to 0)}
}
\value{
imputed_vals
}
\description{
pmm_impute
}
