% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{simulation}
\alias{simulation}
\title{simulation}
\usage{
simulation(
  data,
  simu_times = 500,
  mice_method = FALSE,
  m = 20,
  bayeswt = TRUE,
  donors = 5,
  matchtype = 1,
  ridge = 1e-05
)
}
\arguments{
\item{simu_times}{Number of simulations to run}

\item{ridge}{}
}
\value{
q_simu
}
\description{
simulation
}
