% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{single_impute}
\alias{single_impute}
\title{single_impute}
\usage{
single_impute(
  data,
  mice_method = FALSE,
  m = 20,
  bayeswt = TRUE,
  donors = 5,
  matchtype = 1,
  ridge = 1e-05
)
}
\arguments{
\item{data}{A dataset generated by either create_norm_miss_data() or create_skewed_miss_data()}

\item{mice_method}{Imputation function to use -> mice.impute.pmm() (TRUE) or self-designed function (FALSE)}

\item{m}{Number of imputations to implement}

\item{bayeswt}{Type of regression weights -> "Bayesian" (TRUE) or "Bootstrap" (FALSE)}

\item{donors}{The size of donor pool (usually between 3 and 10)}

\item{matchtype}{Type of matching distance -> {0, 1, 2}}

\item{ridge}{Ridge penalty (a small number close to 0)}
}
\value{
results
}
\description{
single_impute
}
